:truck: Rename: 패키지명 변경
<footer>
- 관련: #230
