+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/common/converter/PublisherTypeConverter.java
@@ -1,11 +1,11 @@
 package com.namo.spring.db.mysql.common.converter;
 
-import com.namo.spring.db.mysql.domains.alarm.type.PublisherType;
+import com.namo.spring.db.mysql.domains.notification.type.PublisherType;
 
 public class PublisherTypeConverter extends AbstractEnumAttributeConverter<PublisherType> {
-	private static final String ENUM_NAME = "알림 발행자 타입";
+    private static final String ENUM_NAME = "알림 발행자 타입";
 
-	public PublisherTypeConverter() {
-		super(PublisherType.class, false, ENUM_NAME);
-	}
+    public PublisherTypeConverter() {
+        super(PublisherType.class, false, ENUM_NAME);
+    }
 }
