+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/common/converter/ReceiverDeviceTypeConverter.java
@@ -1,11 +1,11 @@
 package com.namo.spring.db.mysql.common.converter;
 
-import com.namo.spring.db.mysql.domains.alarm.type.ReceiverDeviceType;
+import com.namo.spring.db.mysql.domains.notification.type.ReceiverDeviceType;
 
 public class ReceiverDeviceTypeConverter extends AbstractEnumAttributeConverter<ReceiverDeviceType> {
-	private static final String ENUM_NAME = "수신자 디바이스 타입";
+    private static final String ENUM_NAME = "수신자 디바이스 타입";
 
-	public ReceiverDeviceTypeConverter() {
-		super(ReceiverDeviceType.class, false, ENUM_NAME);
-	}
+    public ReceiverDeviceTypeConverter() {
+        super(ReceiverDeviceType.class, false, ENUM_NAME);
+    }
 }
